issue-14: change float to int

+++ models/hotel.go
@@ -11,7 +11,7 @@ type Hotel struct {
 	CreatedAt  time.Time `json:"created_at" db:"created_at"`
 	UpdatedAt  time.Time `json:"updated_at" db:"updated_at"`
 	BlockCode  string    `json:"block_code" db:"block_code"`
-	RoomRate   float64   `json:"room_rate" db:"room_rate"`
+	RoomRate   int       `json:"room_rate" db:"room_rate"`
 	PhotoURL   string    `json:"photo_url" db:"photo_url"`
 	LocationID int       `json:"location_id" db:"location_id"`
 	ContactID  int       `json:"contact_id" db:"contact_id"`
